LPS-183536 Implements new validators
