+++ modules/apps/fragment/fragment-entry-processor/fragment-entry-processor-drop-zone/src/main/java/com/liferay/fragment/entry/processor/drop/zone/DropZoneFragmentEntryProcessor.java
@@ -14,7 +14,6 @@
 
 package com.liferay.fragment.entry.processor.drop.zone;
 
-import com.liferay.fragment.exception.FragmentEntryContentException;
 import com.liferay.fragment.model.FragmentEntryLink;
 import com.liferay.fragment.processor.FragmentEntryProcessor;
 import com.liferay.fragment.processor.FragmentEntryProcessorContext;
@@ -29,18 +28,13 @@ import com.liferay.layout.util.structure.LayoutStructureItem;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONUtil;
-import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.util.ListUtil;
-import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -225,38 +219,6 @@ public class DropZoneFragmentEntryProcessor implements FragmentEntryProcessor {
 		return bodyElement.html();
 	}
 
-	@Override
-	public void validateFragmentEntryHTML(String html, String configuration)
-		throws PortalException {
-
-		Document document = _getDocument(html);
-
-		Elements elements = document.select("lfr-drop-zone");
-
-		if (elements.isEmpty()) {
-			return;
-		}
-
-		Set<String> elementDropZoneIds = new LinkedHashSet<>();
-
-		for (Element element : elements) {
-			String dropZoneId = element.attr("data-lfr-drop-zone-id");
-
-			if (!Validator.isBlank(dropZoneId)) {
-				elementDropZoneIds.add(dropZoneId);
-			}
-		}
-
-		if (!elementDropZoneIds.isEmpty() &&
-			(elementDropZoneIds.size() != elements.size())) {
-
-			throw new FragmentEntryContentException(
-				_language.get(
-					_portal.getResourceBundle(LocaleUtil.getDefault()),
-					"you-must-define-a-unique-id-for-each-drop-zone"));
-		}
-	}
-
 	private Document _getDocument(String html) {
 		Document document = Jsoup.parseBodyFragment(html);
 
@@ -273,13 +235,7 @@ public class DropZoneFragmentEntryProcessor implements FragmentEntryProcessor {
 	private FragmentDropZoneRenderer _fragmentDropZoneRenderer;
 
 	@Reference
-	private Language _language;
-
-	@Reference
 	private LayoutPageTemplateStructureLocalService
 		_layoutPageTemplateStructureLocalService;
-
-	@Reference
-	private Portal _portal;
 
 }
