+++ modules/apps/fragment/fragment-entry-processor/fragment-entry-processor-drop-zone/src/main/java/com/liferay/fragment/entry/processor/drop/zone/DropZoneFragmentEntryValidator.java
@@ -1,0 +1,95 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.fragment.entry.processor.drop.zone;
+
+import com.liferay.fragment.exception.FragmentEntryContentException;
+import com.liferay.fragment.processor.FragmentEntryValidator;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.language.Language;
+import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.Validator;
+
+import java.util.LinkedHashSet;
+import java.util.Set;
+
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
+import org.jsoup.select.Elements;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Eudaldo Alonso
+ */
+@Component(
+	property = "fragment.entry.processor.priority:Integer=6",
+	service = FragmentEntryValidator.class
+)
+public class DropZoneFragmentEntryValidator implements FragmentEntryValidator {
+
+	@Override
+	public void validateFragmentEntryHTML(String html, String configuration)
+		throws PortalException {
+
+		Document document = _getDocument(html);
+
+		Elements elements = document.select("lfr-drop-zone");
+
+		if (elements.isEmpty()) {
+			return;
+		}
+
+		Set<String> elementDropZoneIds = new LinkedHashSet<>();
+
+		for (Element element : elements) {
+			String dropZoneId = element.attr("data-lfr-drop-zone-id");
+
+			if (!Validator.isBlank(dropZoneId)) {
+				elementDropZoneIds.add(dropZoneId);
+			}
+		}
+
+		if (!elementDropZoneIds.isEmpty() &&
+			(elementDropZoneIds.size() != elements.size())) {
+
+			throw new FragmentEntryContentException(
+				_language.get(
+					_portal.getResourceBundle(LocaleUtil.getDefault()),
+					"you-must-define-a-unique-id-for-each-drop-zone"));
+		}
+	}
+
+	private Document _getDocument(String html) {
+		Document document = Jsoup.parseBodyFragment(html);
+
+		Document.OutputSettings outputSettings = new Document.OutputSettings();
+
+		outputSettings.prettyPrint(false);
+
+		document.outputSettings(outputSettings);
+
+		return document;
+	}
+
+	@Reference
+	private Language _language;
+
+	@Reference
+	private Portal _portal;
+
+}
