+++ modules/apps/fragment/fragment-entry-processor/fragment-entry-processor-freemarker/src/main/java/com/liferay/fragment/entry/processor/freemarker/FreeMarkerFragmentEntryProcessor.java
@@ -19,7 +19,6 @@ import com.liferay.fragment.entry.processor.freemarker.internal.configuration.Fr
 import com.liferay.fragment.exception.FragmentEntryContentException;
 import com.liferay.fragment.helper.FragmentEntryLinkHelper;
 import com.liferay.fragment.input.template.parser.FragmentEntryInputTemplateNodeContextHelper;
-import com.liferay.fragment.input.template.parser.InputTemplateNode;
 import com.liferay.fragment.model.FragmentEntryLink;
 import com.liferay.fragment.processor.FragmentEntryProcessor;
 import com.liferay.fragment.processor.FragmentEntryProcessorContext;
@@ -30,19 +29,13 @@ import com.liferay.info.item.InfoItemReference;
 import com.liferay.info.item.InfoItemServiceRegistry;
 import com.liferay.info.item.provider.InfoItemObjectProvider;
 import com.liferay.item.selector.ItemSelector;
-import com.liferay.petra.io.DummyWriter;
 import com.liferay.petra.io.unsync.UnsyncStringWriter;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.module.configuration.ConfigurationProvider;
-import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.servlet.DummyHttpServletResponse;
 import com.liferay.portal.kernel.template.StringTemplateResource;
 import com.liferay.portal.kernel.template.Template;
 import com.liferay.portal.kernel.template.TemplateConstants;
@@ -54,12 +47,10 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
 
 import java.util.ResourceBundle;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -197,84 +188,6 @@ public class FreeMarkerFragmentEntryProcessor
 		return unsyncStringWriter.toString();
 	}
 
-	@Override
-	public void validateFragmentEntryHTML(String html, String configuration)
-		throws PortalException {
-
-		FreeMarkerFragmentEntryProcessorConfiguration
-			freeMarkerFragmentEntryProcessorConfiguration =
-				_configurationProvider.getCompanyConfiguration(
-					FreeMarkerFragmentEntryProcessorConfiguration.class,
-					CompanyThreadLocal.getCompanyId());
-
-		if (!freeMarkerFragmentEntryProcessorConfiguration.enable() ||
-			!_isFreemarkerTemplate(html)) {
-
-			return;
-		}
-
-		Template template = TemplateManagerUtil.getTemplate(
-			TemplateConstants.LANG_TYPE_FTL,
-			new StringTemplateResource("template_id", "[#ftl] " + html), true);
-
-		try {
-			HttpServletRequest httpServletRequest = null;
-			HttpServletResponse httpServletResponse = null;
-
-			ServiceContext serviceContext =
-				ServiceContextThreadLocal.getServiceContext();
-
-			if (serviceContext != null) {
-				httpServletRequest = serviceContext.getRequest();
-				httpServletResponse = serviceContext.getResponse();
-			}
-
-			if (httpServletResponse == null) {
-				httpServletResponse = new DummyHttpServletResponse();
-			}
-
-			if ((httpServletRequest != null) &&
-				(httpServletRequest.getAttribute(WebKeys.THEME_DISPLAY) !=
-					null)) {
-
-				JSONObject configurationDefaultValuesJSONObject =
-					_fragmentEntryConfigurationParser.
-						getConfigurationDefaultValuesJSONObject(configuration);
-
-				template.putAll(
-					HashMapBuilder.<String, Object>put(
-						"configuration", configurationDefaultValuesJSONObject
-					).put(
-						"fragmentElementId", StringPool.BLANK
-					).put(
-						"fragmentEntryLinkNamespace", StringPool.BLANK
-					).put(
-						"input",
-						new InputTemplateNode(
-							StringPool.BLANK, StringPool.BLANK,
-							StringPool.BLANK, "name", false, false, false,
-							"type", "value")
-					).put(
-						"layoutMode", Constants.VIEW
-					).putAll(
-						_fragmentEntryConfigurationParser.getContextObjects(
-							configurationDefaultValuesJSONObject, configuration,
-							null, new long[0])
-					).build());
-
-				template.prepareTaglib(httpServletRequest, httpServletResponse);
-
-				template.prepare(httpServletRequest);
-
-				template.processTemplate(new DummyWriter());
-			}
-		}
-		catch (TemplateException templateException) {
-			throw new FragmentEntryContentException(
-				_getMessage(templateException), templateException);
-		}
-	}
-
 	private Object _getInfoItem(InfoItemReference infoItemReference) {
 		if (infoItemReference == null) {
 			return null;
@@ -322,14 +235,6 @@ public class FreeMarkerFragmentEntryProcessor
 		}
 
 		return message;
-	}
-
-	private boolean _isFreemarkerTemplate(String html) {
-		if (html.contains("${") || html.contains("<#") || html.contains("<@")) {
-			return true;
-		}
-
-		return false;
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
