+++ modules/apps/fragment/fragment-entry-processor/fragment-entry-processor-resources/src/main/java/com/liferay/fragment/entry/processor/resources/ResourcesFragmentEntryProcessor.java
@@ -41,8 +41,4 @@ public class ResourcesFragmentEntryProcessor implements FragmentEntryProcessor {
 			fragmentEntryLink, html);
 	}
 
-	@Override
-	public void validateFragmentEntryHTML(String html, String configuration) {
-	}
-
 }
