+++ modules/apps/fragment/fragment-test/src/testIntegration/java/com/liferay/fragment/contributor/test/FragmentCollectionContributorPropagationTest.java
@@ -417,11 +417,6 @@ public class FragmentCollectionContributorPropagationTest {
 			return html;
 		}
 
-		@Override
-		public void validateFragmentEntryHTML(String html, String configuration)
-			throws PortalException {
-		}
-
 		private final Map<Long, String> _companyIdsMap = new HashMap<>();
 		private final String _fragmentEntryKey;
